pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
# TODO(rename): change this to your module name (kebab-case, prefix with `pmscope-`)
name = "pmscope-mymodule"
version = "0.0.1"
# TODO(describe): one-sentence description of what your module does
description = "PM-SCOPE module template — fork and customize"
requires-python = ">=3.11"
dependencies = [
    # pmscope-sdk is installed from the vendored wheel in wheels/.
    # See README.md "Vendored wheel" section for why and how to refresh it.
    "pmscope-sdk==0.1.0",
    "uvicorn~=0.27.0",
    # TODO(deps): add your tool-specific dependencies here (e.g. shodan, requests, etc.)
]

[project.optional-dependencies]
testing = ["pytest~=7.4.4", "pytest-asyncio~=0.23.3", "httpx~=0.26.0"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
